+++ .gitignore
@@ -1,6 +1,6 @@
 # Personal
 /materials
-/.vscode
+.vscode/
 
 # Created by https://www.toptal.com/developers/gitignore/api/macos,windows,visualstudiocode,python,django
 # Edit at https://www.toptal.com/developers/gitignore?templates=macos,windows,visualstudiocode,python,django
