--- articles/serializers.py
@@ -1,12 +1,0 @@
-from rest_framework.serializers import ModelSerializer
-from .models import Article
-
-
-class ArticleSerializer(ModelSerializer):
-    class Meta:
-        model = Article
-        fields = (
-            "id",
-            "title",
-            "content",
-        )
