+++ requirements.txt
@@ -1,16 +1,6 @@
-asgiref==3.6.0
-black==23.3.0
-click==8.1.3
 Django==4.2
 django-cors-headers==3.14.0
 django-rest-framework==0.1.0
 djangorestframework==3.14.0
 djangorestframework-simplejwt==5.2.2
-mypy-extensions==1.0.0
-packaging==23.1
-pathspec==0.11.1
-platformdirs==3.2.0
-PyJWT==2.6.0
-pytz==2023.3
-sqlparse==0.4.4
-tomli==2.0.1
+python-dotenv==1.0.0
