+++ static/css/styles.css
@@ -1,15 +1,51 @@
+* {
+  box-sizing: border-box;
+}
+
+a {
+  user-select: none;
+  cursor: pointer;
+}
+
+form {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin: 0;
+}
+
+form button {
+  width: 150px;
+}
+
 .title {
   text-align: center;
 }
-
-ul {
-  list-style: none;
+.btns {
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
 
-.article {
-  width: 100%;
-  padding: 0 24px;
-  margin-bottom: 8px;
-  border: 1px solid black;
-  border-radius: 4px;
+.btns > a {
+  text-decoration-line: none;
+  width: 150px;
+  height: 60px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 8px 0;
+  margin: 0 4px;
+  border-radius: 6px;
+  text-align: center;
+  font-size: 16px;
+  font-weight: bold;
+  color: white;
+  background-color: #118bee;
+}
+
+.form-wrapper {
+  width: 60%;
+  margin: 0 auto;
 }
