+++ templates/home.html
@@ -7,6 +7,13 @@
 {%block content%}
 <div id="container">
   <h1 class="title">홈 화면</h1>
+  <div class="btns">
+    <a href="{%url 'login'%}">로그인</a>
+    <a href="{%url 'signup'%}">회원가입</a>
+  </div>
+  <div class="btns">
+    <button id="logout-btn" type="button">로그아웃</button>
+  </div>
 </div>
 
 <script src="{% static 'js/index.js' %}"></script>
